requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
flatbuffers==25.2.10
fonttools==4.59.1
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jax==0.7.0
jaxlib==0.7.0
kiwisolver==1.4.9
matplotlib==3.10.5
mediapipe==0.10.14
ml_dtypes==0.5.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
scipy==1.16.1
setuptools==80.9.0
setuptools-scm==8.3.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.3
